image/test/mochitest/test_bug553982.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=553982
-->
<head>
  <title>Test for Bug 553982</title>
  <script type="application/javascript" src="/MochiKit/MochiKit.js"></script>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=553982">Mozilla Bug 553982</a>
<pre id="test">
<script type="application/javascript">

SimpleTest.waitForExplicitFinish();
var success = false;

// This should fire
function doError() {
  success = true;
}

// This should not fire
function doLoad() {
  ok(false, "onload should not fire");
}

window.onload = function() {
  ok(success, "onerror should fire");
  SimpleTest.finish();
}

</script>
</pre>
<img src="data:text/html," onerror="doError()" onload="doLoad()" />
</body>
</html>
